// src/forgetrustee/frontend/src/types/generated/metadataFieldOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MetadataOptionAdditionalDetailsDto } from './metadataOptionAdditionalDetailsDto';

export interface MetadataFieldOption {
  /** The underlying value (machine-readable) */
  value: string;
  /** The human-readable label to show in the UI */
  label: string;
  /** Additional detail input configuration for this option */
  additionalDetails?: MetadataOptionAdditionalDetailsDto;
  /** Marker indicating that the frontend may render this option in a visually separated extra area */
  extraArea?: boolean;
  /** If true, additional detail inputs should be hidden/disabled for this option */
  disableAdditionalDetails?: boolean;
}