// src/forgetrustee/frontend/src/types/generated/metadataFieldDtoType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * The type of the field
 */
export type MetadataFieldDtoType = typeof MetadataFieldDtoType[keyof typeof MetadataFieldDtoType];


export const MetadataFieldDtoType = {
  TEXT: 'TEXT',
  LONGTEXT: 'LONGTEXT',
  SELECT: 'SELECT',
  TAGS: 'TAGS',
  TAGS_WITH_DESCRIPTION_LIST: 'TAGS_WITH_DESCRIPTION_LIST',
  TAGS_WITH_TWO_FIELDS: 'TAGS_WITH_TWO_FIELDS',
  TAGS_WITH_TEXT_AND_DESCRIPTION_LIST: 'TAGS_WITH_TEXT_AND_DESCRIPTION_LIST',
  TEXT_WITH_DESCRIPTION: 'TEXT_WITH_DESCRIPTION',
  TEXT_WITH_DESCRIPTION_LIST: 'TEXT_WITH_DESCRIPTION_LIST',
} as const;